+++ package.json
@@ -6,10 +6,11 @@
   "scripts": {
     "clean": "rm -fR dist/",
     "tsc": "tsc",
+    "tsc-paths": "tsc-alias",
     "copy-public": "cp -R ./src/public/ dist/public",
     "copy-views": "cp -R ./src/views/ dist/views",
-    "build": "npm run clean && npm run tsc && npm run copy-public && npm run copy-views",
-    "dev": "ts-node-dev src/index.ts"
+    "build": "npm run clean && npm run tsc && npm run tsc-paths && npm run copy-public && npm run copy-views",
+    "dev": "ts-node-dev -r tsconfig-paths/register src/index.ts"
   },
   "keywords": [],
   "author": "",
@@ -24,6 +25,8 @@
     "eslint-plugin-n": "^15.3.0",
     "eslint-plugin-promise": "^6.0.1",
     "ts-node-dev": "^2.0.0",
+    "tsc-alias": "^1.7.0",
+    "tsconfig-paths": "^4.1.0",
     "typescript": "^4.8.3"
   },
   "dependencies": {
