+++ src/core/test/models/article.entity.ts
