+++ src/index.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata';
-import app from './app';
+import app from './bootstrap';
 
 app.listen(app.get('port'), () => {
     console.log('Server on port:', app.get('port'));
